refactor:moved quorum election starting mechanism to service layer
